# api_doc.yaml
openapi: 3.0.3
info:
    title: meta_tech
    version: 1.0.0
    contact: {}
    description: |-
        - api for shop. 
        - (admin account: email = admin@gmail.com, password = 123aaaaa)
servers:
    - url: 'http://localhost:4000'
paths:
...
    /api/item:
        get:
            tags:
                - Item
            summary: get list item
            description: get list item in shop
            operationId: getListItem
            parameters:
                - name: minPrice
                  in: query
                  schema:
                      type: string
                      example: '11000'
                - name: maxPrice
                  in: query
                  schema:
                      type: string
                      example: '11000'
                - name: category
                  in: query
                  schema:
                      type: string
                      example: '9'
                - name: limit
                  in: query
                  schema:
                      type: string
                      example: '1'
                - name: page
                  in: query
                  schema:
                      type: string
                      example: '1'
                - name: search
                  in: query
                  schema:
                      type: string
                      example: a
                - name: isDesc
                  in: query
                  schema:
                      type: string
                      example: '1'

            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
...
    /api/item/add:
        post:
            tags:
                - Item
            summary: add item
            description: add item
            operationId: addItem
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                description:
                                    type: string
                                    example: >-
                                        Time to elevate your T-shirt collection? This tee answers
                                        the call. Heavyweight cotton and a clean, embroidered Futura
                                        logo keep it simple for pairing up with any outfit. The
                                        oversize fit gives you plenty of extra space for casual
                                        style and easy layering.
                                id_category:
                                    type: string
                                    example: '9'
                                image:
                                    type: string
                                    format: binary
                                import_price:
                                    type: string
                                    example: '60000'
                                name:
                                    type: string
                                    example: 'Mũ len classic '
                                price:
                                    type: string
                                    example: '89000'
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
...
    /api/item/delete/{id_item}:
        delete:
            tags:
                - Item
            summary: delete item
            description: delete item
            operationId: deleteItem
            parameters:
                - name: id_item
                  example: 1451
                  in: path
                  description: id of item
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: ''
    /api/cart/add:
        post:
            tags:
                - Cart
            summary: add to cart
            description: add color item to cart
            operationId: addToCart
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            type: object
                            properties:
                                id_color_item:
                                    type: string
                                    example: '7133'
                                quantity:
                                    type: string
                                    example: '2'
                        examples:
                            add to cart:
                                value:
                                    id_color_item: '7133'
                                    quantity: '2'
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
...
    /api/cart/delete/{id_cartItem}:
        delete:
            tags:
                - Cart
            summary: delete cart item
            description: delete cart's item
            operationId: deleteCartItem
            parameters:
                - name: id_cartItem
                  example: 23
                  in: path
                  description: id of cart's item
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
...
    /api/order/cancel/{id_order}:
        put:
            tags:
                - Order For Custommer
            summary: cancel order
            description: cancel order
            operationId: cancelOrder
            parameters:
                - name: id_order
                  example: 23
                  in: path
                  description: id of order
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
...
    /api/rating/add:
        post:
            tags:
                - Rating
            summary: add rating
            description: add rating
            operationId: addRating
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            type: object
                            properties:
                                comment:
                                    type: string
                                    example: >-
                                        good product for user !! but the delivering service is quite
                                        bad :)
                                id_item:
                                    type: string
                                    example: '1438'
                                id_order_item:
                                    type: string
                                    example: '22'
                                star:
                                    type: string
                                    example: '4'
                        examples:
                            add rating:
                                value:
                                    comment: >-
                                        good product for user !! but the delivering service is quite
                                        bad :)
                                    id_item: '1438'
                                    id_order_item: '22'
                                    star: '4'
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
    /api/rating/delete/{id_rating}:
        delete:
            tags:
                - Rating
            summary: delete rating
            description: delete rating
            operationId: deleteRating
            parameters:
                - name: id_rating
                  example: 23
                  in: path
                  description: id of rating
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
    /api/rating/update/{id_rating}:
        put:
            tags:
                - Rating
            summary: update rating
            description: update rating
            operationId: updateRating
            parameters:
                - name: id_rating
                  example: 23
                  in: path
                  description: id of rating
                  required: true
                  schema:
                      type: integer
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            type: object
                            properties:
                                comment:
                                    type: string
                                    example: rất xinh đẹp
                                star:
                                    type: string
                                    example: '4'
                        examples:
                            update rating:
                                value:
                                    comment: rất xinh đẹp
                                    star: '4'
            responses:
                '200':
                    description: ''
            security:
                - bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
...